% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLloyd.R
\name{ARI}
\alias{ARI}
\title{Calculate ARI of two sets of labels.}
\usage{
ARI(z, z.hat, mode = c("averaged", "minimal"))
}
\arguments{
\item{z}{First clustering vector.}

\item{z.hat}{Second clustering vector.}

\item{mode}{Averaged/Minimal ARI.}
}
\value{
Calculated ARI.
}
\description{
Calculate ARI of two sets of labels.
}
