% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLloyd.R
\name{HO.Lloyd}
\alias{HO.Lloyd}
\title{High-order Lloyd Algorithm.}
\usage{
HO.Lloyd(Y, z, t_max = 10)
}
\arguments{
\item{Y}{Observed tensor.}

\item{z}{Vector of initialized labels.}

\item{t_max}{maximum number of iterations.}
}
\value{
Estimated clustering labels.
}
\description{
High-order Lloyd Algorithm.
}
