% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLloyd.R
\name{HO.SC}
\alias{HO.SC}
\title{High-order tensor spectral clustering.}
\usage{
HO.SC(Y, r)
}
\arguments{
\item{Y}{Observed tensor.}

\item{r}{Vector of cluster numbers.}
}
\value{
clustering labels.
}
\description{
High-order tensor spectral clustering.
}
