% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLloyd.R
\name{SC}
\alias{SC}
\title{Matricized tensor spectral clustering using kmeans++.}
\usage{
SC(Y, r)
}
\arguments{
\item{Y}{Observed tensor.}

\item{r}{Vector of cluster numbers.}
}
\value{
clustering labels.
}
\description{
Matricized tensor spectral clustering using kmeans++.
}
