% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLloyd.R
\name{tomembershipmatrix}
\alias{tomembershipmatrix}
\title{Transform a label vector to a membership matrix.
Non-exported.}
\usage{
tomembershipmatrix(z)
}
\arguments{
\item{z}{vector of clustering labels.}
}
\value{
Corresponding membership matrix.
}
\description{
Transform a label vector to a membership matrix.
Non-exported.
}
