% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLloyd.R
\name{toweightmatrix}
\alias{toweightmatrix}
\title{Transform a label vector to a weighted matrix
Non-exported.}
\usage{
toweightmatrix(z)
}
\arguments{
\item{z}{vector of clustering labels}
}
\value{
Corresponding weighted matrix
}
\description{
Transform a label vector to a weighted matrix
Non-exported.
}
